Refactor tests and add lots of more tests
